# [class-parse] Use `BufferedStream` for up to 10x performance gain