--- a/src/Xamarin.Android.Tools.Bytecode/ClassPath.cs
+++ b/src/Xamarin.Android.Tools.Bytecode/ClassPath.cs
@@ -72,7 +72,8 @@ namespace Xamarin.Android.Tools.Bytecode {
 						if (!ClassFile.IsClassFile (s) || entry.Name.EndsWith (".jnilib", StringComparison.OrdinalIgnoreCase))
 							continue;
 					}
-					using (var s = entry.Open ()) {
+					using (var entry_stream = entry.Open ())
+					using (var s = new BufferedStream (entry_stream)) {
 						try {
 							var c   = new ClassFile (s);
 							Add (c);